+++ public/pages/Changelog.md
@@ -2,6 +2,8 @@
 
 Changelog for the QuickDoc React versions.
 
+## v0.1.7
+
 ## v0.1.6
 
 ### Features